requirements.txt
asgiref==3.8.1
Django==5.1.6
djangorestframework==3.15.2
djangorestframework-simplejwt>=4.7.2
psycopg==3.2.4
psycopg-binary==3.2.4
sqlparse==0.5.3
